+++ pytest/test_bem_potential_velocity_and_free_surface_elevation.py
@@ -219,6 +219,10 @@ def test_compute_free_surface_elevation_on_free_surface(solver, result):
 #                            Check values                             #
 #######################################################################
 
+def test_pressure_integration(solver, result):
+    f = result.body.integrate_pressure(solver.compute_pressure(result.body.mesh, result))
+    assert f == result.forces
+
 def test_reconstruction_of_given_boundary_condition(solver, result):
     velocities = solver.compute_velocity(result.body.mesh, result)
     normal_velocities = np.einsum('...k,...k->...', velocities, result.body.mesh.faces_normals)